# FE_and_BE_SWP391_SE2039_G06_SportsCenterManagement/SWP391- SportsCenterManagement- frontend - Copy/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sports Center Management System (SCMS) - RESTful API Specification",
    "description": "Tài liệu đặc tả chuẩn RESTful API cho Hệ thống Quản lý Trung tâm Thể thao SCMS (Sports Center Management System).\n\nĐồng bộ 100% với mã nguồn Frontend và các cập nhật nghiệp vụ mới nhất (02/10/2026):\n- Bổ nhiệm nhân sự từ tài khoản có sẵn (Giới hạn vai trò COACH / RECEPTIONIST, cấm cấp MANAGER)\n- Giáo án chung cho lớp học & Giáo án riêng cá nhân hóa cho học viên (Cập nhật, hủy giáo án riêng quay về giáo án chung)\n- Tích hợp 15 môn thể thao Olympic, 9 cụm cơ sở vật chất, 4 nhóm vai trò (Manager, Receptionist, Coach, Member).",
    "version": "1.1.0",
    "contact": {
      "name": "SCMS Development Team (SE2039_G06)",
      "email": "support@scms.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Spring Boot Backend Server (Default REST API)"
    },
    {
      "url": "http://localhost:3003/api",
      "description": "Vite Dev Server (Frontend Reverse Proxy)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập chuỗi JWT Token xác thực nhận được từ endpoint `/auth/login`. Header định dạng: `Authorization: Bearer <token>`"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr-mgr-01"
          },
          "email": {
            "type": "string",
            "example": "manager@scms.vn"
          },
          "fullName": {
            "type": "string",
            "example": "Trần Văn Quản Lý"
          },
          "phone": {
            "type": "string",
            "example": "0987111222"
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGER",
              "RECEPTIONIST",
              "COACH",
              "MEMBER"
            ],
            "example": "MANAGER"
          },
          "memberCode": {
            "type": "string",
            "example": "MEM-5621"
          },
          "specialty": {
            "type": "string",
            "example": "Bơi lội & Gym thể hình"
          },
          "certification": {
            "type": "string",
            "example": "AFC Level A & NASM-CPT"
          },
          "packageId": {
            "type": "string",
            "example": "pkg-all-access"
          },
          "packageName": {
            "type": "string",
            "example": "Gói All-Access Olympic Pass"
          },
          "packageStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "EXPIRED"
            ],
            "example": "ACTIVE"
          },
          "packageExpiry": {
            "type": "string",
            "example": "2026-12-31"
          },
          "avatar": {
            "type": "string",
            "example": "https://images.unsplash.com/photo-1534528741775-53994a69daeb?w=150"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-15"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "manager@scms.vn"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "fullName",
          "email",
          "phone",
          "password"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn Mới"
          },
          "email": {
            "type": "string",
            "example": "nguyenvanmoi@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "0912999888"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        }
      },
...
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "email",
          "otp",
          "newPassword"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "member@scms.vn"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          },
          "newPassword": {
            "type": "string",
            "example": "newpassword123"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Nội dung thông báo lỗi chi tiết từ máy chủ"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "1. Authentication"
        ],
        "summary": "Đăng nhập tài khoản bằng Email + Mật khẩu",
        "description": "Xác thực tài khoản và trả về JWT Bearer Token kèm thông tin người dùng (Role: MANAGER, RECEPTIONIST, COACH, MEMBER).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công, trả về JWT Token và User Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email không tồn tại hoặc sai mật khẩu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Tài khoản bị tạm khóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "1. Authentication"
        ],
        "summary": "Đăng ký tài khoản hội viên mới",
        "description": "Tạo tài khoản hội viên (Role: MEMBER) với mã thẻ sinh tự động (MEM-XXXX).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký tài khoản thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Email đã được sử dụng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "tags": [
          "1. Authentication"
        ],
        "summary": "Khôi phục mật khẩu thông qua mã OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Khôi phục mật khẩu thành công"
          },
          "400": {
            "description": "Mã OTP không hợp lệ hoặc email không tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}